Move get_sync_db_engine from dependencies to utils
