+++ app/dependencies.py
@@ -15,7 +15,6 @@ import redis
 from fastapi import BackgroundTasks, Depends, HTTPException, Request, status
 from jose import jwt
 from pydantic import ValidationError
-from sqlalchemy.engine import Engine, create_engine
 from sqlalchemy.ext.asyncio import (
     AsyncEngine,
     AsyncSession,
@@ -55,19 +54,6 @@ async def get_request_id(request: Request) -> str:
     The request identifier is a unique UUID which is used to associate logs saved during the same request
     """
     return request.state.request_id
-
-
-def get_sync_db_engine(settings: Settings) -> Engine:
-    """
-    Create a synchronous database engine
-    """
-    if settings.SQLITE_DB:
-        SQLALCHEMY_DATABASE_URL = f"sqlite:///./{settings.SQLITE_DB}"
-    else:
-        SQLALCHEMY_DATABASE_URL = f"postgresql://{settings.POSTGRES_USER}:{settings.POSTGRES_PASSWORD}@{settings.POSTGRES_HOST}/{settings.POSTGRES_DB}"
-
-    engine = create_engine(SQLALCHEMY_DATABASE_URL, echo=settings.DATABASE_DEBUG)
-    return engine
 
 
 def get_db_engine(settings: Settings) -> AsyncEngine:
